refactoring LogInFragment.kt
