+++ app/src/main/java/com/example/lecture1/ui/fragments/LogInFragment.kt
@@ -7,6 +7,7 @@ import androidx.fragment.app.Fragment
 import android.view.View
 import android.widget.Button
 import android.widget.Toast
+import androidx.core.widget.doAfterTextChanged
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import by.kirich1409.viewbindingdelegate.viewBinding
@@ -63,25 +64,9 @@ class LogInFragment : Fragment(R.layout.fragment_log_in) {
             viewModel.login(email = inputEmail?.text.toString(), password = inputPassword?.text.toString())
         }
 
-        inputEmail?.addTextChangedListener(object : TextWatcher {
-            override fun afterTextChanged(s: Editable?) { }
+        inputEmail?.doAfterTextChanged { viewModel.isValidate(email = inputEmail?.text.toString(), password = inputPassword?.text.toString()) }
 
-            override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) { }
-
-            override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
-                viewModel.isValidate(email = inputEmail?.text.toString(), password = inputPassword?.text.toString())
-            }
-        })
-
-        inputPassword?.addTextChangedListener(object : TextWatcher {
-            override fun afterTextChanged(s: Editable?) { }
-
-            override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) { }
-
-            override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
-                viewModel.isValidate(email = inputEmail?.text.toString(), password = inputPassword?.text.toString())
-            }
-        })
+        inputPassword?.doAfterTextChanged { viewModel.isValidate(email = inputEmail?.text.toString(), password = inputPassword?.text.toString()) }
     }
 
     override fun onDestroyView() {
