+++ app/src/main/java/com/example/lecture1/ui/fragments/OnBoardingFragment.kt
@@ -21,6 +21,5 @@ class OnBoardingFragment : Fragment(R.layout.fragment_on_boarding) {
             val action = OnBoardingFragmentDirections.actionOnBoardingFragmentToLogInFragment()
             findNavController().navigate(action)
         }
-
     }
 }
